fix: extract grouped table logic to groupedTableOperations.tsx
